test/10: Moved tests to match general folder structure
