+++ Streetcode/Streetcode.XUnitTest/BLL/MediatR/News/GetNewsByIdTests.cs
@@ -1,4 +1,5 @@
-﻿using AutoMapper;
+﻿using System.Linq.Expressions;
+using AutoMapper;
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore.Query;
 using Moq;
@@ -6,24 +7,24 @@ using Streetcode.BLL.DTO.Media.Images;
 using Streetcode.BLL.DTO.News;
 using Streetcode.BLL.Interfaces.BlobStorage;
 using Streetcode.BLL.Interfaces.Logging;
-using Streetcode.BLL.MediatR.Newss.GetByUrl;
+using Streetcode.BLL.MediatR.Newss.GetById;
 using Streetcode.DAL.Repositories.Interfaces.Base;
 using Streetcode.DAL.Repositories.Interfaces.Newss;
-using System.Linq.Expressions;
 using Xunit;
 
-namespace Streetcode.XUnitTest.BLL_Tests.MediatR.News;
+namespace Streetcode.XUnitTest.BLL.MediatR.News;
 
-public class GetNewsByUrlTests
+public class GetNewsByIdTests
 {
+
     private readonly Mock<IRepositoryWrapper> _mockRepositoryWrapper;
     private readonly Mock<INewsRepository> _mockNewsRepository;
     private readonly Mock<IMapper> _mockMapper;
     private readonly Mock<IBlobService> _mockBlobService;
     private readonly Mock<ILoggerService> _mockLogger;
-    private readonly GetNewsByUrlHandler _handler;
+    private readonly GetNewsByIdHandler _handler;
 
-    public GetNewsByUrlTests()
+    public GetNewsByIdTests()
     {
         _mockRepositoryWrapper = new Mock<IRepositoryWrapper>();
         _mockNewsRepository = new Mock<INewsRepository>();
@@ -33,7 +34,7 @@ public class GetNewsByUrlTests
 
         _mockRepositoryWrapper.Setup(x => x.NewsRepository).Returns(_mockNewsRepository.Object);
 
-        _handler = new GetNewsByUrlHandler(
+        _handler = new GetNewsByIdHandler(
             _mockMapper.Object,
             _mockRepositoryWrapper.Object,
             _mockBlobService.Object,
@@ -41,13 +42,12 @@ public class GetNewsByUrlTests
     }
 
     [Fact]
-    public async Task GetNewsByUrl_WhenNewsExists_ShouldReturnDTO()
+    public async Task GetNewsById_WhenNewsExists_ShouldReturnDTO()
     {
         // Arrange
-        string newsUrl = "test-news-url";
         int newsId = 1;
-        var newsEntity = CreateNewsEntity(newsId, newsUrl);
-        var newsDTO = CreateNewsDTO(newsId, newsUrl);
+        var newsEntity = CreateNewsEntity(newsId);
+        var newsDTO = CreateNewsDTO(newsId);
 
         _mockNewsRepository
             .Setup(r => r.GetFirstOrDefaultAsync(
@@ -59,7 +59,7 @@ public class GetNewsByUrlTests
             .Setup(m => m.Map<NewsDTO>(newsEntity))
             .Returns(newsDTO);
 
-        var query = new GetNewsByUrlQuery(newsUrl);
+        var query = new GetNewsByIdQuery(newsId);
 
         // Act
         var result = await _handler.Handle(query, CancellationToken.None);
@@ -68,28 +68,29 @@ public class GetNewsByUrlTests
         result.Should().NotBeNull();
         result.IsSuccess.Should().BeTrue();
         result.Value.Should().BeEquivalentTo(newsDTO);
-        result.Value.URL.Should().Be(newsUrl);
+        result.Value.Id.Should().Be(newsId);
 
         _mockMapper.Verify(m => m.Map<NewsDTO>(newsEntity), Times.Once);
     }
 
     [Fact]
-    public async Task GetNewsByUrl_WhenNewsNotFound_ShouldReturnFailure()
+    public async Task GetNewsById_WhenNewsNotFound_ShouldReturnFailure()
     {
         // Arrange
-        string newsUrl = "non-existing-url";
+        int newsId = 1;
+        var newsEntity = CreateNewsEntity(newsId);
 
         _mockNewsRepository
             .Setup(r => r.GetFirstOrDefaultAsync(
                 It.IsAny<Expression<Func<DAL.Entities.News.News, bool>>>(),
                 It.IsAny<Func<IQueryable<DAL.Entities.News.News>, IIncludableQueryable<DAL.Entities.News.News, object>>>()))
-            .ReturnsAsync((DAL.Entities.News.News)null);
+            .ReturnsAsync(newsEntity);
 
         _mockMapper
-            .Setup(m => m.Map<NewsDTO>(null))
+            .Setup(m => m.Map<NewsDTO>(newsEntity))
             .Returns((NewsDTO)null);
 
-        var query = new GetNewsByUrlQuery(newsUrl);
+        var query = new GetNewsByIdQuery(newsId);
 
         // Act
         var result = await _handler.Handle(query, CancellationToken.None);
@@ -99,19 +100,16 @@ public class GetNewsByUrlTests
         result.IsSuccess.Should().BeFalse();
         result.Errors.Should().NotBeEmpty();
 
-
-        _mockMapper.Verify(m => m.Map<NewsDTO>(null), Times.Once);
+        _mockMapper.Verify(m => m.Map<NewsDTO>(newsEntity), Times.Once);
     }
 
-
     [Fact]
-    public async Task GetNewsByUrl_WhenNewsHasImage_ShouldSetBase64ForImage()
+    public async Task GetNewsById_WhenNewsHasImage_ShouldSetBase64ForImage()
     {
         // Arrange
-        string newsUrl = "news-with-image";
         int newsId = 1;
-        var newsEntity = CreateNewsEntity(newsId, newsUrl);
-        var newsDTO = CreateNewsDTOWithImage(newsId, "test-image.jpg", newsUrl);
+        var newsEntity = CreateNewsEntity(newsId);
+        var newsDTO = CreateNewsDTOWithImage(newsId, "test-image.jpg");
 
         _mockNewsRepository
             .Setup(r => r.GetFirstOrDefaultAsync(
@@ -127,7 +125,7 @@ public class GetNewsByUrlTests
             .Setup(b => b.FindFileInStorageAsBase64("test-image.jpg"))
             .Returns("base64-encoded-image-data");
 
-        var query = new GetNewsByUrlQuery(newsUrl);
+        var query = new GetNewsByIdQuery(newsId);
 
         // Act
         var result = await _handler.Handle(query, CancellationToken.None);
@@ -138,47 +136,46 @@ public class GetNewsByUrlTests
         result.Value.Image.Should().NotBeNull();
         result.Value.Image.Base64.Should().Be("base64-encoded-image-data");
         result.Value.Image.BlobName.Should().Be("test-image.jpg");
-        result.Value.URL.Should().Be(newsUrl);
 
         _mockBlobService.Verify(b => b.FindFileInStorageAsBase64("test-image.jpg"), Times.Once);
         _mockMapper.Verify(m => m.Map<NewsDTO>(newsEntity), Times.Once);
     }
 
-    private DAL.Entities.News.News CreateNewsEntity(int id, string url = null)
+    private DAL.Entities.News.News CreateNewsEntity(int id)
     {
         return new DAL.Entities.News.News
         {
             Id = id,
             Title = $"Test News {id}",
             Text = $"Test Text {id}",
-            URL = url ?? $"test-news-{id}",
+            URL = $"test-news-{id}",
             CreationDate = DateTime.UtcNow,
             ImageId = null,
             Image = null
         };
     }
 
-    private NewsDTO CreateNewsDTO(int id, string url = null)
+    private NewsDTO CreateNewsDTO(int id)
     {
         return new NewsDTO
         {
             Id = id,
             Title = $"Test News {id}",
             Text = $"Test Text {id}",
-            URL = url ?? $"test-news-{id}",
+            URL = $"test-news-{id}",
             CreationDate = DateTime.UtcNow,
             Image = null
         };
     }
 
-    private NewsDTO CreateNewsDTOWithImage(int id, string blobName, string url = null)
+    private NewsDTO CreateNewsDTOWithImage(int id, string blobName)
     {
         return new NewsDTO
         {
             Id = id,
             Title = $"Test News {id}",
             Text = $"Test Text {id}",
-            URL = url ?? $"test-news-{id}",
+            URL = $"test-news-{id}",
             CreationDate = DateTime.UtcNow,
             Image = new ImageDTO
             {
