+++ Streetcode/Streetcode.XUnitTest/BLL/MediatR/News/UpdateNewsTests.cs
@@ -12,7 +12,7 @@ using Streetcode.DAL.Repositories.Interfaces.Base;
 using Streetcode.DAL.Repositories.Interfaces.Newss;
 using Xunit;
 
-namespace Streetcode.XUnitTest.BLL_Tests.MediatR.News;
+namespace Streetcode.XUnitTest.BLL.MediatR.News;
 
 public class UpdateNewsTests
 {
